requirements.txt
anyio==4.4.0
APScheduler==3.6.3
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
cachetools==4.2.2
cairocffi==1.7.1
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
cssselect2==0.7.0
decorator==5.1.1
diff-match-patch==20230430
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
django-import-export==4.1.1
django-jet==1.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.53.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.2
Markdown==3.6
oscrypto==1.3.0
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.3.1
pyphen==0.15.0
pypng==0.20220715.0
python-bidi==0.5.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.4
python-telegram-bot-calendar==1.0.5
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
rfc3986==1.5.0
setuptools==70.3.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
svglib==1.5.1
tablib==3.5.0
tinycss2==1.3.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3
